creating the first test for App.test.js
